fix: Improve retrieving Quests Widget performances - MEED-9358 - Meeds-io/meeds#3449 (#1933)
Prior to this change, when retrieving quests from Store, the
`expand=expandPrerequisites` was used while it's not necessary anymore.
This change will delete this expand option in order to reduce the size
of retrieved response from REST endpoint and reduces the computation
time in server end when displaying the Quests Widget.

Resolves Meeds-io/meeds#3449