Fix: Remove the access to go back in program detail page - MEED-9582 - meeds-io/meeds#3419 (#1958) (#1959)
Before this change, from a program detail page, it was possible to go back to the list using a button (arrow-left + program name are clickable), Program detail page can be accessed using the drawer and it is no more use to go back to the programs list, URL changes but it doesn't work. To fix this problem, remove the arrow-left button and change replaceState to pushState for browser history. After this change, the arrow-left and the click to the program name are deleted and the user can close the drawer or go back using the browser options.
Resolves https://github.com/Meeds-io/meeds/issues/3419